pyproject.toml
[project]
name = "test-task-ioka"
version = "0.1.0"
description = ""
authors = [
    {name = "OPELsinus",email = "92687099+OPELsinus@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "passlib (==1.7.4)",
    "bcrypt (==4.0.1)",
    "black (>=25.1.0,<26.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
